server/pyproject.toml
[project]
name = "curalink"
version = "1.0.0"
description = "Curalink AI Medical Research Assistant — FastAPI Pipeline"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "groq==0.11.0",
    "sentence-transformers==3.1.1",
    "motor==3.6.0",
    "pymongo==4.10.1",
    "httpx==0.27.2",
    "python-dotenv==1.0.1",
    "langsmith==0.1.147",
    "pydantic==2.9.2",
    "numpy==1.26.4",
    "scikit-learn==1.5.2",
    "python-multipart==0.0.12",
    "aiohttp==3.10.10",
    "xmltodict==0.13.0",
    "tenacity==9.0.0",
    "slowapi==0.1.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
